fix image upload and get route + using get route to display cocktail in homepage
