+++ backend/src/Controllers/image.controller.ts
@@ -5,10 +5,25 @@ import {
   Query,
   Res,
   UploadedFile,
+  UseInterceptors,
 } from '@nestjs/common';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 import { ImageService } from 'src/Services/image.service';
 import path = require('path');
+import { diskStorage } from 'multer';
+import { FileInterceptor } from '@nestjs/platform-express';
+
+const storage = {
+  storage: diskStorage({
+    destination: './uploads/cocktails',
+    filename: (_, file, cb) => {
+      const filename: string = path.parse(file.originalname).name;
+      const extension: string = path.parse(file.originalname).ext;
+
+      cb(null, `${filename}${extension}`);
+    },
+  }),
+};
 
 @ApiTags('Images')
 @Controller('images')
@@ -16,8 +31,9 @@ export class ImageController {
   constructor(private readonly imageService: ImageService) {}
 
   @Post('/upload')
+  @UseInterceptors(FileInterceptor('file', storage))
   upload(@UploadedFile() file) {
-    return this.imageService.upload(file);
+    return { file: file.filename };
   }
 
   @ApiResponse({
