+++ frontend/src/styles/home.css
@@ -96,6 +96,7 @@
     height: 200px;
     border-radius: 20px 20px 0 0;
     margin: -20px;
+    object-fit: cover;
 }
 
 .card-footer {
